+++ src/scripts/domManipulation.js
@@ -1,8 +1,7 @@
 import * as generator from './domTools';
 import { retrieveItem, saveItem } from './localStorage';
-import { setDate, getDate, hideShowDropdown } from './viewProjectTasks';
-import { todoList } from './classes/todoListItem';
 import * as hui from './handlingUserInput';
+import createSingleTask from './createSingleTask';
 
 const mainContainer = generator.htmlGenerator('div', 'todo-list-tasks', 'todoListTasks');
 
@@ -50,162 +49,11 @@ const todoListMainContainer = () => {
     return mainContainer;
   };
 
-  const dropDownBuilder = (id) => {
-    const container = generator.htmlGenerator('div', 'drop-container', `dropContainer${id}`);
-    container.classList.add('hidden');
-    container.classList.add('slide-in-bottom');
-    const dropdownQuery = retrieveItem('dropdownState');
-    if (dropdownQuery[1] === id) {
-      if (dropdownQuery[0] === false) {
-        container.classList.remove('hidden');
-        container.classList.remove('slide-in-bottom');
-      }
-    }
-
-    const descriptionSection = () => {
-      const descriptionContainer = generator.htmlGenerator('div', 'description-container', `descriptionContainer${id}`);
-      const descriptionInput = generator.htmlGenerator('input', 'description-input', `descriptionInput${id}`);
-
-      if (hui.getTaskProperty(id, 'description')) {
-        descriptionInput.placeholder = `${hui.getTaskProperty(id, 'description')}`;
-      } else {
-        descriptionInput.placeholder = 'Give your task a description...';
-      }
-
-      const descriptionSubmit = generator.htmlGenerator('button', 'description-submit-button', `descriptionSubmitButton${id}`);
-      const descriptionSubmitIcon = generator.textGenerator('i', '<i class="fas fa-save"></i>');
-      descriptionSubmit.addEventListener('click', () => {
-        hui.setTaskProperty('descriptionInput', id, 'description');
-        generator.reload();
-      });
-      descriptionSubmit.classList.add('hidden');
-      generator.enterShortcut(descriptionSubmit, descriptionInput);
-
-      descriptionSubmit.appendChild(descriptionSubmitIcon);
-      descriptionContainer.append(descriptionInput, descriptionSubmit);
-      return descriptionContainer;
-    };
-
-    const dateSection = () => {
-      const dateContainer = generator.htmlGenerator('div', 'date-container', `dateContainer${id}`);
-      const dateInput = generator.htmlGenerator('input', 'date-input', `dateInput${id}`);
-      dateInput.type = 'datetime-local';
-      dateInput.setAttribute('min', '1066-01-01T00:00');
-      dateInput.setAttribute('max', '2100-01-01T00:00');
-      dateInput.classList.add('hidden');
-
-      const dateSubmit = generator.htmlGenerator('button', 'date-submit-button', `dateSubmitButton${id}`);
-
-      const dateSubmitIcon = generator.textGenerator('i', '<i class="fas fa-check-circle"></i>');
-      dateSubmit.appendChild(dateSubmitIcon);
-      dateSubmit.classList.add('hidden');
-      const dateBackButton = generator.htmlGenerator('button', 'date-back-button', `dateBackButton${id}`);
-      dateBackButton.innerHTML = 'Cancel';
-
-      const dateDisplay = generator.htmlGenerator('input', 'date-display', `dateDisplay${id}`);
-      dateDisplay.placeholder = getDate(id);
-
-      dateDisplay.addEventListener('click', (event) => {
-        event.preventDefault();
-        generator.hideAndShow(dateDisplay, dateInput, dateBackButton, dateSubmit);
-      });
-
-      dateBackButton.addEventListener('click', (event) => {
-        event.preventDefault();
-        generator.hideAndShow(dateInput, dateDisplay, dateBackButton);
-      });
-      dateBackButton.classList.add('hidden');
-      dateSubmit.addEventListener('click', (event) => {
-        event.preventDefault();
-        setDate(id);
-        generator.hideAndShow(dateInput, dateDisplay, dateBackButton);
-        generator.reload();
-      });
-      generator.enterShortcut(dateSubmit, dateInput);
-
-      dateContainer.append(dateInput, dateDisplay, dateSubmit, dateBackButton);
-      return dateContainer;
-    };
-
-    container.append(descriptionSection(), dateSection());
-    return container;
-  };
-
-  const createSingleTask = (id) => {
-    const savedItems = currentProject.items;
-    const savedItemsLength = savedItems.length;
-    const simpleListItemContainer = generator.htmlGenerator('div', 'simple-list-item-container', `simpleListItemContainer${id}`);
-    const listItemContainer = generator.htmlGenerator('div', 'todo-list-item-container', `listItemContainer${id}`);
-    const listItemInputContainer = generator.htmlGenerator('input', 'project-task-input', `projectTask${id}`);
-    const listItemSubmitButton = generator.htmlGenerator('button', 'project-task-submit', `projectTaskSubmit${id}`);
-    const listItemSubmitIcon = generator.textGenerator('i', '<i class="fas fa-check-circle"></i>');
-    const listItemDeleteButtonText = generator.textGenerator('p', '<i class="fas fa-times"></i>');
-    const listItemDeleteButton = generator.htmlGenerator('button', 'list-item-delete-button', `projectTaskDelete${id}`);
-    const listItemPriorityButton = generator.htmlGenerator('button', 'todo-list-item-button', `listItemButton${id}`);
-    const listItemDownArrow = generator.htmlGenerator('button', 'list-item-down-arrow', `listItemDownArrow${id}`);
-    const listItemDownArrowIcon = generator.textGenerator('p', '<i class="fas fa-sort-down"></i>');
-    listItemInputContainer.placeholder = 'Give it a title...';
-    listItemDownArrow.appendChild(listItemDownArrowIcon);
-    listItemSubmitButton.appendChild(listItemSubmitIcon);
-    listItemPriorityButton.classList.add('list-item-priority');
-    listItemDownArrow.addEventListener('click', () => {
-      hideShowDropdown(id);
-    });
-
-    if (id && currentProject.items[id]) {
-      setTimeout(() => {
-        if (currentProject.items[id].priority === 0) {
-          listItemPriorityButton.innerHTML = '<i class="fas fa-star"></i>';
-        } else {
-          listItemPriorityButton.innerHTML = '<i class="far fa-star"></i>';
-        }
-      }, 1);
-    }
-
-    const setPriorityStatus = () => {
-      if (hui.settingPriority() === 0) {
-        listItemPriorityButton.innerHTML = '<i class="fas fa-star"></i>';
-      } else {
-        listItemPriorityButton.innerHTML = '<i class="far fa-star"></i>';
-      }
-    };
-
-    listItemPriorityButton.addEventListener('click', setPriorityStatus);
-
-
-    listItemSubmitButton.classList.add('hidden');
-
-    listItemDeleteButton.type = 'button';
-    listItemDeleteButton.appendChild(listItemDeleteButtonText);
-
-    const listItemID = generator.generateID(listItemInputContainer);
-    if (listItemID >= savedItemsLength) {
-      listItemSubmitButton.addEventListener('click', hui.saveTask);
-    } else {
-      listItemSubmitButton.addEventListener('click', hui.editTask);
-    }
-
-    generator.enterShortcut(listItemSubmitButton, listItemInputContainer);
-
-    listItemDeleteButton.addEventListener('click', (event) => {
-      event.preventDefault();
-      hui.obliterateTask();
-    });
-    listItemContainer.append(
-      listItemInputContainer,
-      listItemSubmitButton,
-      listItemPriorityButton,
-      listItemDeleteButton,
-      listItemDownArrow,
-    );
-    simpleListItemContainer.append(listItemContainer, dropDownBuilder(id));
-    return simpleListItemContainer;
-  };
 
   const todoListTasks = () => {
     if (currentProject) {
       for (let i = 0; i < currentProject.items.length; i++) {
-        mainContainer.appendChild(createSingleTask(`${i}`));
+        mainContainer.appendChild(createSingleTask(`${i}`, currentProject));
         setTimeout(() => {
           const inputField = document.getElementById(`projectTask${i}`);
           if (currentProject.items[i].title) {
@@ -223,11 +71,11 @@ const todoListMainContainer = () => {
     const domLength = listContainer.childNodes.length;
     if (currentProject) {
       if (listLength >= 0 && domLength <= listLength) {
-        listContainer.appendChild(createSingleTask(listLength));
+        listContainer.appendChild(createSingleTask(listLength, currentProject));
       } else if (listLength >= 0 && domLength <= listLength + 1) {
         listContainer.innerHTML = '';
         todoListTasks();
-        listContainer.appendChild(createSingleTask(listLength));
+        listContainer.appendChild(createSingleTask(listLength, currentProject));
       }
     }
   };
